electron trigger sf, make EGM implementation new default custom implementation moved to Run2
