+++ pocket_coffea/lib/scale_factors.py
@@ -59,55 +59,11 @@ def get_ele_sf(
             ak.unflatten(sfup, counts),
             ak.unflatten(sfdown, counts),
         )
-    elif key == 'trigger':
-        electron_correctionset = correctionlib.CorrectionSet.from_file(
-            electronSF.trigger_sf[year]["file"]
-        )
-        map_name = electronSF.trigger_sf[year]["name"]
-
-        output = {}
-        trigger_path = electronSF.trigger_sf[year]["path"]
-        for variation in variations:
-            if variation == "nominal":
-                output[variation] = [
-                    electron_correctionset[map_name].evaluate(
-                        year_pog,
-                        "sf",
-                        trigger_path,
-                        eta.to_numpy(),
-                        pt.to_numpy(),
-                    )
-                ]
-            else:
-                # Nominal sf==1
-                nominal = np.ones_like(pt.to_numpy())
-                # Systematic variations
-                output[variation] = [
-                    nominal,
-                    electron_correctionset[map_name].evaluate(
-                        year_pog,
-                        f"{variation}up",
-                        trigger_path,
-                        eta.to_numpy(),
-                        pt.to_numpy(),
-                    ),
-                    electron_correctionset[map_name].evaluate(
-                        year_pog,
-                        f"{variation}down",
-                        electronSF.trigger_sf[year]["path"],
-                        eta.to_numpy(),
-                        pt.to_numpy(),
-                    ),
-                ]
-            for i, sf in enumerate(output[variation]):
-                output[variation][i] = ak.unflatten(sf, counts)
-
-        return output
     else:
-        raise Exception(f"Invalid key `{key}` for get_ele_sf. Available keys are 'reco', 'id', 'trigger'.")
-    
+        raise Exception(f"Invalid key `{key}` for get_ele_sf. Available keys are 'reco', 'id'.")
 
-def sf_ele_trigger_EGM(params, events, year):
+
+def sf_ele_trigger(params, events, year):
     """Compute electron trigger scale factors using the EGM JSON files with correctionlib.
     Returns the per-event scale factor for the trigger.
 
@@ -263,39 +219,6 @@ def sf_ele_id(params, events, year):
 
     # The SF arrays corresponding to the electrons are multiplied along the electron axis in order to obtain a per-event scale factor.
     return ak.prod(sf_id, axis=1), ak.prod(sfup_id, axis=1), ak.prod(sfdown_id, axis=1)
-
-
-def sf_ele_trigger(params, events, year, variations=["nominal"]):
-    '''
-    This function computes the semileptonic electron trigger SF by considering the leading electron in the event.
-    This computation is valid only in the case of the semileptonic final state.
-    Additionally, also the up and down variations of the SF for a set of systematic uncertainties are returned.
-    '''
-    coll = params.lepton_scale_factors.electron_sf.collection
-    ele_pt = events[coll].pt
-    ele_eta = events[coll].etaSC
-
-    ele_pt_flat, ele_eta_flat, ele_counts = (
-        ak.flatten(ele_pt),
-        ak.flatten(ele_eta),
-        ak.num(ele_pt),
-    )
-    sf_dict = get_ele_sf(
-        params,
-        year,
-        pt=ele_pt_flat,
-        eta=ele_eta_flat,
-        phi=None,
-        counts=ele_counts,
-        key='trigger',
-        variations=variations,
-    )
-
-    for variation in sf_dict.keys():
-        for i, sf in enumerate(sf_dict[variation]):
-            sf_dict[variation][i] = ak.prod(sf, axis=1)
-
-    return sf_dict
 
 
 def sf_mu(params, events, year, key=''):
