+++ pocket_coffea/lib/weights/common/common.py
@@ -4,7 +4,6 @@ import numpy as np
 from pocket_coffea.lib.scale_factors import (
     sf_ele_reco,
     sf_ele_id,
-    sf_ele_trigger,
     sf_mu,
     sf_btag,
     sf_btag_calib,
@@ -99,41 +98,6 @@ SF_L1prefiring = WeightLambda.wrap_func(
     has_variations=True
     )
 
-
-########################################################################
-# More complicated WeightWrapper defining dynamic variations depending
-# on the data taking period
-
-class SF_ele_trigger(WeightWrapper):
-    name = "sf_ele_trigger"
-    has_variations = True
-
-    def __init__(self, params, metadata):
-        super().__init__(params, metadata)
-        # Getting the variations from the parameters depending on the year
-        self._variations = params.systematic_variations.weight_variations.sf_ele_trigger[metadata["year"]]
-
-    def compute(self, events, size, shape_variation):
-        if shape_variation == "nominal":
-            out = sf_ele_trigger(self._params, events, self._metadata["year"],
-                             variations= ["nominal"] + self._variations,
-                             )
-            # This is a dict with variation: [nom, up, down]
-            return WeightDataMultiVariation(
-                 name = self.name,
-                 nominal = out["nominal"][0],
-                 variations = self._variations,
-                 up = [out[var][1] for var in self._variations],
-                 down = [out[var][2] for var in self._variations]
-            )
-
-        else:
-            out = sf_ele_trigger(self._params, events, self._metadata["year"],
-                             variations= ["nominal"] )
-            return WeightData(
-                name = self.name,
-                nominal = out["nominal"][0]
-            )
         
 ########################################
 # Btag scale factors have weights depending on the shape_variation
@@ -324,7 +288,6 @@ common_weights = [
     SF_ele_id,
     SF_mu_id,
     SF_mu_iso,
-    SF_ele_trigger,
     SF_mu_trigger,
     SF_btag,
     SF_btag_calib,
