+++ pocket_coffea/lib/weights/common/weights_run3.py
@@ -1,11 +1,11 @@
-from ..weights import WeightWrapper, WeightLambda, WeightData, WeightDataMultiVariation
-from pocket_coffea.lib.scale_factors import sf_ele_trigger_EGM
+from pocket_coffea.lib.scale_factors import sf_ele_trigger
 
-
+from ..weights import WeightData, WeightDataMultiVariation, WeightLambda, WeightWrapper
 
 SF_ele_trigger = WeightLambda.wrap_func(
-    name="sf_ele_trigger_egm",
-    function=lambda params, metadata, events, size, shape_variations:
-        sf_ele_trigger_EGM(params, events, metadata["year"]),
-    has_variations=True
-    )
+    name="sf_ele_trigger",
+    function=lambda params, metadata, events, size, shape_variations: sf_ele_trigger(
+        params, events, metadata["year"]
+    ),
+    has_variations=True,
+)
