+++ pocket_coffea/parameters/lepton_scale_factors.yaml
@@ -111,6 +111,23 @@ lepton_scale_factors:
       '2018':
         name: ???
         file: ???
+      2022_preEE:
+        2022_preEE:
+        name: Electron-HLT-SF
+        file: /cvmfs/cms.cern.ch/rsync/cms-nanoAOD/jsonpog-integration/POG/EGM/2022_Summer22/electronHlt.json.gz
+        path: ???
+      2022_postEE:
+        name: Electron-HLT-SF
+        file: /cvmfs/cms.cern.ch/rsync/cms-nanoAOD/jsonpog-integration/POG/EGM/2022_Summer22EE/electronHlt.json.gz
+        path: ???
+      2023_preBPix:
+        name: Electron-HLT-SF
+        file: /cvmfs/cms.cern.ch/rsync/cms-nanoAOD/jsonpog-integration/POG/EGM/2023_Summer23/electronHlt.json.gz
+        path: ???
+      2023_postBPix:
+        name: Electron-HLT-SF
+        file: /cvmfs/cms.cern.ch/rsync/cms-nanoAOD/jsonpog-integration/POG/EGM/2023_Summer23BPix/electronHlt.json.gz
+        path: ???
 
   ##################################################################################################  
   muon_sf:
